# fix: Nested serialize.Executor.AsyncExec calls may deadlock when overwhelmed 