Signed-off-by: Denis Tingajkin <denis.tingajkin@xored.com>

## Motivation

https://github.com/networkservicemesh/sdk/issues/512