@denis-tingajkin 2020-10-13
Bench test results

## Before

[root]goos: darwin
goarch: amd64
pkg: github.com/networkservicemesh/sdk/pkg/tools/serialize
BenchmarkExecutorSync
BenchmarkExecutorSync-8   	 1908020	       616 ns/op
PASS


goos: darwin
goarch: amd64
pkg: github.com/networkservicemesh/sdk/pkg/tools/serialize
BenchmarkExecutorSync
BenchmarkExecutorSync-8   	 2145922	       544 ns/op
PASS

## After

```
goos: darwin
goarch: amd64
pkg: github.com/networkservicemesh/sdk/pkg/tools/serialize
BenchmarkExecutorAsync
BenchmarkExecutorAsync-8   	 4798452	       236 ns/op
PASS

goos: darwin
goarch: amd64
pkg: github.com/networkservicemesh/sdk/pkg/tools/serialize
BenchmarkExecutorSync
BenchmarkExecutorSync-8   	 1908020	       616 ns/op
PASS
```