@denis-tingajkin 2020-10-14
@edwarnicke I also did unboundedchannel but it wors x3 slow than this version and currently, I did not see the problems with mutex variant