@edwarnicke 2020-10-15
Checkout this: https://github.com/edwarnicke/serialize/pull/3