@denis-tingajkin 2020-10-16
@edwarnicke Lets move SDK to use executor from the separate repo :)