@denis-tingajkin 2020-10-16
Will be fixed by https://github.com/edwarnicke/serialize/pull/3#pullrequestreview-510447947